dabom-backend/src/main/java/joeuncamp/dabombackend/domain/wish/service/WishService.java
package joeuncamp.dabombackend.domain.wish.service;

import joeuncamp.dabombackend.domain.course.repository.CourseJpaRepository;
import joeuncamp.dabombackend.domain.member.repository.MemberJpaRepository;
import joeuncamp.dabombackend.domain.wish.dto.WishRequestDto;
import joeuncamp.dabombackend.domain.wish.repository.WishJpaRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class WishService {
    private final MemberJpaRepository memberJpaRepository;
    private final CourseJpaRepository courseJpaRepository;
    private final WishJpaRepository wishJpaRepository;

    public void toggleWish(WishRequestDto wishRequestDto){

    }
}
